add new step for invest
